Rename file example -> bandsplot in js/
