+++ js/lib/bandsplot.js
