+++ js/lib/index.js
@@ -1,3 +1,3 @@
 // Export widget models and views, and the npm package version number.
-module.exports = require('./example.js');
+module.exports = require('./bandsplot.js');
 module.exports['version'] = require('../package.json').version;
